setting up react router to project & move to Home route
